+++ src/dibujos/index.ts
@@ -1,3 +1,3 @@
 export * from './base'
 export * from './estaciones'
-export * from './cantero'
+export * from './vacios'